+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Small ui fixes
 - Can set TC String and IDConsent seperately now
 - Reworked error handling
+- More documentation
+- Removed unused function
 
 ### Fixed
 - Url schemes error
@@ -46,4 +48,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug when using login flow with permission
 - Fixed a bug when some claims were missing
 - Make get/update permission possible in permission flow
+- Fixed an edge case when resuming a session
+- Fixed app2app redirect